remove types file and refactor AliasDialog to email_aliases_list
